TC-587 flytter domeneklasser ut i egen fil
