+++ src/main/java/no/nav/pto/veilarbportefolje/arbeidssoeker/v2/ArbeidssoekerDomene.kt
@@ -1,5 +1,6 @@
 package no.nav.pto.veilarbportefolje.arbeidssoeker.v2
 
+import no.nav.common.types.identer.Fnr
 import java.time.ZonedDateTime
 import java.util.UUID
 
@@ -13,12 +14,47 @@ data class OpplysningerOmArbeidssoeker(
     val opplysningerOmJobbsituasjon: OpplysningerOmArbeidssoekerJobbsituasjon
 )
 
-
 data class OpplysningerOmArbeidssoekerJobbsituasjon(
     val opplysningerOmArbeidssoekerId: UUID,
     val jobbsituasjon: List<String>
 )
 
+data class ArbeidssoekerPeriode(
+    val arbeidssoekerperiodeId: UUID,
+    val fnr: Fnr
+)
+
+data class Profilering(
+    val periodeId: UUID,
+    val profileringsresultat: Profileringsresultat,
+    val sendtInnTidspunkt: ZonedDateTime
+)
+
+enum class Profileringsresultat {
+    UKJENT_VERDI,
+    UDEFINERT,
+    ANTATT_GODE_MULIGHETER,
+    ANTATT_BEHOV_FOR_VEILEDNING,
+    OPPGITT_HINDRINGER
+}
+
+data class ProfileringResponse(
+    val profileringId: UUID,
+    val periodeId: UUID,
+    val opplysningerOmArbeidssoekerId: UUID,
+    val sendtInnAv: MetadataResponse,
+    val profilertTil: ProfilertTil,
+    val jobbetSammenhengendeSeksAvTolvSisteManeder: Boolean?,
+    val alder: Int?
+)
+
+fun ProfileringResponse.toProfilering(): Profilering {
+    return Profilering(
+        periodeId = this.periodeId,
+        profileringsresultat = Profileringsresultat.valueOf(this.profilertTil.name),
+        sendtInnTidspunkt = this.sendtInnAv.tidspunkt
+    )
+}
 
 fun OpplysningerOmArbeidssoekerResponse.toOpplysningerOmArbeidssoeker() = OpplysningerOmArbeidssoeker(
     opplysningerOmArbeidssoekerId = this.opplysningerOmArbeidssoekerId,
