+++ src/main/java/no/nav/pto/veilarbportefolje/arbeidssoeker/v2/OppslagArbeidssoekerregisteretClient.kt
@@ -103,7 +103,7 @@ data class OpplysningerOmArbeidssoekerResponse(
 
 data class BeskrivelseMedDetaljerResponse(
     val beskrivelse: JobbSituasjonBeskrivelse,
-    val detaljer: Map<String, String>   // TODO: Litt usikker på om typene er riktig for key/value
+    val detaljer: Map<String, String>
 )
 
 enum class JobbSituasjonBeskrivelse {
@@ -140,16 +140,6 @@ data class UtdanningResponse(
 
 // Profilering typer
 data class ProfileringRequest(val identitetsnummer: String, val periodeId: UUID)
-
-data class ProfileringResponse(
-    val profileringId: UUID,
-    val periodeId: UUID,
-    val opplysningerOmArbeidssoekerId: UUID,
-    val sendtInnAv: MetadataResponse,
-    val profilertTil: ProfilertTil,
-    val jobbetSammenhengendeSeksAvTolvSisteManeder: Boolean?,
-    val alder: Int?
-)
 
 enum class ProfilertTil {
     UKJENT_VERDI,
